+++ app/Models/Event.php
@@ -2,6 +2,7 @@
 
 namespace App\Models;
 
+use App\Core\Application;
 use App\Core\DbModel;
 
 abstract class Event extends DbModel
@@ -14,6 +15,7 @@ abstract class Event extends DbModel
     public string $type = '';
     public string $price = '';
     public int $capacity = 0;
+    public int $organiserId;
 
 
     public function getTableName() : string
@@ -28,10 +30,15 @@ abstract class Event extends DbModel
             'picture' => [$this->validator::RULE_REQUIRED],
             'description' => [$this->validator::RULE_REQUIRED],
             'type' => [$this->validator::RULE_REQUIRED],
-            'eventDate' => [$this->validator::RULE_REQUIRED],
+            'eventDate' => [$this->validator::RULE_REQUIRED, [$this->validator::RULE_DATE, 'minDate' => (new \DateTime())->modify('+7 days')->format('Y-m-d')]],
             'categoryId' => [$this->validator::RULE_REQUIRED],
             'capacity' => [$this->validator::RULE_REQUIRED]
         ];
+    }
+
+    public static function getAll() : array
+    {
+        return self::getAll();
     }
 
 }
