+++ app/Views/events/create.html.twig
@@ -200,4 +200,5 @@
 
 {% block scripts %}
     <script src="/assets/js/script.js"></script>
+    <script src="/assets/js/main.js"></script>
 {% endblock %}
