+++ app/Views/layouts/main.html.twig
@@ -100,8 +100,6 @@
 </nav>
 
 {% block content %}{% endblock %}
-
 {% block scripts %}{% endblock %}
-<script src="/assets/js/main.js"></script>
 </body>
 </html>
