+++ app/Views/macros/picture.html.twig
@@ -5,7 +5,7 @@
         <div class="space-y-2">
             <label class="block text-sm font-medium text-gray-300">{{ label }}</label>
             <div class="text-xs text-gray-400">Drag and drop or click to upload</div>
-            <input name="{{ name }}" type="file" class="absolute inset-0 w-full h-full opacity-0 cursor-pointer"/>
+            <input name="{{ name }}" type="file" class="absolute inset-0 w-full h-full opacity-0 cursor-pointer "/>
         </div>
     </div>
     {% if errors[name] is defined %}
