+++ public/assets/css/style.css
@@ -89,3 +89,11 @@ body {
     border: 1px solid #3D3D4D;
     color: white;
 }
+
+.event-card {
+    transition: transform 0.3s ease, box-shadow 0.3s ease;
+}
+.event-card:hover {
+    transform: translateY(-10px);
+    box-shadow: 0 10px 20px rgba(0,0,0,0.2);
+}
