+++ public/assets/js/main.js
@@ -146,3 +146,121 @@ document.addEventListener('DOMContentLoaded', (event) => {
         });
     }
 });
+
+
+// Sample event data
+const events = [
+    {
+        category: 'Music',
+        date: 'June 15, 2024',
+        title: 'Summer Music Festival',
+        location: 'Downtown Central Park',
+        price: '$75',
+        image: 'https://via.placeholder.com/400x250?text=Music+Festival'
+    },
+    {
+        category: 'Technology',
+        date: 'July 22, 2024',
+        title: 'Tech Innovation Summit',
+        location: 'Convention Center',
+        price: '$150',
+        image: 'https://via.placeholder.com/400x250?text=Tech+Summit'
+    },
+    {
+        category: 'Arts',
+        date: 'August 5, 2024',
+        title: 'Contemporary Art Exhibition',
+        location: 'Modern Art Gallery',
+        price: '$40',
+        image: 'https://via.placeholder.com/400x250?text=Art+Exhibition'
+    },
+    {
+        category: 'Sports',
+        date: 'September 12, 2024',
+        title: 'Marathon Running Event',
+        location: 'City Stadium',
+        price: '$60',
+        image: 'https://via.placeholder.com/400x250?text=Marathon+Event'
+    },
+    {
+        category: 'Music',
+        date: 'October 20, 2024',
+        title: 'Jazz Night',
+        location: 'Downtown Jazz Club',
+        price: '$45',
+        image: 'https://via.placeholder.com/400x250?text=Jazz+Night'
+    },
+    {
+        category: 'Technology',
+        date: 'November 8, 2024',
+        title: 'AI and Future Tech Conference',
+        location: 'Tech Center',
+        price: '$200',
+        image: 'https://via.placeholder.com/400x250?text=AI+Conference'
+    }
+];
+
+// Pagination variables
+let currentPage = 1;
+const eventsPerPage = 3;
+const totalPages = Math.ceil(events.length / eventsPerPage);
+
+// Function to render events
+function renderEvents(page) {
+    const eventGrid = document.getElementById('eventGrid');
+    eventGrid.innerHTML = ''; // Clear previous events
+
+    const start = (page - 1) * eventsPerPage;
+    const end = start + eventsPerPage;
+    const pageEvents = events.slice(start, end);
+
+    pageEvents.forEach(event => {
+        const eventCard = `
+                    <div class="event-card bg-gray-800 rounded-xl overflow-hidden border border-gray-700">
+                        <img 
+                            src="${event.image}" 
+                            alt="Event Image" 
+                            class="w-full h-56 object-cover"
+                        >
+                        <div class="p-6">
+                            <div class="flex justify-between items-center mb-2">
+                                <span class="text-orange-500 font-medium">${event.category}</span>
+                                <span class="text-gray-400">${event.date}</span>
+                            </div>
+                            <h3 class="text-xl font-bold mb-2">${event.title}</h3>
+                            <p class="text-gray-400 mb-4">${event.location}</p>
+                            <div class="flex justify-between items-center">
+                                <span class="text-lg font-bold text-orange-500">${event.price}</span>
+                                <button class="bg-gray-700 text-white px-4 py-2 rounded-lg hover:bg-gray-600">
+                                    View Details
+                                </button>
+                            </div>
+                        </div>
+                    </div>
+                `;
+        eventGrid.innerHTML += eventCard;
+    });
+
+    // Update pagination buttons
+    document.getElementById('currentPage').textContent = `Page ${page}`;
+    document.getElementById('prevPage').disabled = page === 1;
+    document.getElementById('nextPage').disabled = page === totalPages;
+}
+
+// Event listeners for pagination
+document.getElementById('prevPage').addEventListener('click', () => {
+    if (currentPage > 1) {
+        currentPage--;
+        renderEvents(currentPage);
+    }
+});
+
+document.getElementById('nextPage').addEventListener('click', () => {
+    if (currentPage < totalPages) {
+        currentPage++;
+        renderEvents(currentPage);
+    }
+});
+
+// Initial render
+renderEvents(currentPage);
