+++ public/index.php
@@ -3,7 +3,6 @@
 require_once __DIR__ . '/../vendor/autoload.php';
 
 use App\Core\Application;
-
 $config = [];
 
 $config['db'] = require_once __DIR__ . '/../app/config/database.php';
@@ -14,4 +13,3 @@ require_once __DIR__ . '/../app/core/web.php';
 $app->db->applyMigrations();
 
 $app->run();
-
